# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AMIDDOCS OFFICIAL SWAGGER",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/students/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "create student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Create Student",
                "parameters": [
                    {
                        "description": "create student dto",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/studentmodel.CreateStudentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studentmodel.StudentDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/students/get-by-id": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "get student by id from query param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "GetStudentByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "student id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studentmodel.StudentDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/students/info": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "get student by id from query param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "GetStudentByID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studentmodel.StudentDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/get-by-id": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "return user by id from path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Return User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "get user by id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usermodel.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/info": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "return user by id from path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Return User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usermodel.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "login by login and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "login form",
                        "name": "loginForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usermodel.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tokenmodel.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/refresh-token": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "refresh token by old token and user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "refresh token",
                "parameters": [
                    {
                        "description": "refresh dto with user id and old token",
                        "name": "refreshdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rtokenmodel.RefreshDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tokenmodel.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "amiderrors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "depmodel.DepartmentDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "groupfields.EducationForm": {
            "type": "string",
            "enum": [
                "FULL_TIME",
                "EXTRAMURAL"
            ],
            "x-enum-varnames": [
                "FULL_TIME",
                "EXTRAMURAL"
            ]
        },
        "groupmodel.GroupDTO": {
            "type": "object",
            "properties": {
                "educationFinishDate": {
                    "type": "string"
                },
                "educationForm": {
                    "$ref": "#/definitions/groupfields.EducationForm"
                },
                "educationStartDate": {
                    "type": "string"
                },
                "educationYear": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isBudget": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "studyDepartmentId": {
                    "type": "integer"
                }
            }
        },
...
        "rtokenmodel.RefreshDTO": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "stdocmodel.StudentDocumentDTO": {
            "type": "object",
            "properties": {
                "docNumber": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "orderDate": {
                    "type": "string"
                },
                "orderNumber": {
                    "type": "string"
                },
                "studentId": {
                    "type": "integer"
                },
                "studyStartDate": {
                    "type": "string"
                }
            }
        },
        "studentmodel.CreateStudentDTO": {
            "type": "object",
            "properties": {
                "docNumber": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fatherName": {
                    "type": "string"
                },
                "groupName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orderDate": {
                    "type": "string"
                },
                "orderNumber": {
                    "type": "string"
                },
                "studyStartDate": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "studentmodel.StudentDTO": {
            "type": "object",
            "properties": {
                "department": {
                    "$ref": "#/definitions/depmodel.DepartmentDTO"
                },
                "document": {
                    "$ref": "#/definitions/stdocmodel.StudentDocumentDTO"
                },
                "group": {
                    "$ref": "#/definitions/groupmodel.GroupDTO"
                },
                "id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/usermodel.UserDTO"
                }
            }
        },
        "tokenmodel.TokenResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userfields.Role"
                    }
                }
            }
        },
        "userfields.Role": {
            "type": "string",
            "enum": [
                "STUDENT",
                "ADMIN",
                "SECRETARY"
            ],
            "x-enum-varnames": [
                "STUDENT",
                "ADMIN",
                "SECRETARY"
            ]
        },
...
        "usermodel.LoginForm": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "usermodel.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fatherName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userfields.Role"
                    }
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer Token Auth",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Token": {
            "description": "Client Token auth",
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}